handle payment loading, error and success states
handle /bin/zsh orders in holiday programs and play lab
add receipt to holiday programs confirmation email
