+++ server/src/sendgrid/types.ts
@@ -5,6 +5,7 @@ export type Emails = {
         location: string
         address: string
         bookings: { datetime: string; confirmationPage: string }[]
+        receiptUrl: string | undefined
     }
 
     kingsvilleOpeningConfirmation: {
@@ -26,7 +27,7 @@ export type Emails = {
         location: string
         bookings: { time: string; details: string; confirmationPage: string }[]
         isTermEnrolment: boolean
-        receiptUrl?: string
+        receiptUrl: string | undefined
     }
 
     // AFTER SCHOOL PROGRAM
